Group extension tests into one file
